requirements.txt
agentlightning[apo]
openai
pyyaml
python-dotenv
